use Eq typeclass in Assertion - fixes #87
