+++ src/main/scala/com/github/agourlay/cornichon/http/HttpAssertions.scala
@@ -8,7 +8,7 @@ import com.github.agourlay.cornichon.http.server.HttpMockServerResource.SessionK
 import com.github.agourlay.cornichon.json.JsonAssertions.JsonAssertion
 import com.github.agourlay.cornichon.resolver.Resolver
 import com.github.agourlay.cornichon.steps.regular.assertStep.{ AssertStep, CustomMessageAssertion, GenericAssertion }
-import com.github.agourlay.cornichon.util.ShowInstances._
+import com.github.agourlay.cornichon.util.Instances._
 
 object HttpAssertions {
 
@@ -27,8 +27,8 @@ object HttpAssertions {
     def is(expected: (String, String)*) = from_session_step(
       title = s"headers is ${displayStringPairs(expected)}",
       key = SessionKey(lastResponseHeadersKey),
-      expected = s ⇒ expected.map { case (name, value) ⇒ s"$name$headersKeyValueDelim$value" },
-      mapValue = (session, sessionHeaders) ⇒ sessionHeaders.split(",").toSeq
+      expected = s ⇒ expected.toList.map { case (name, value) ⇒ s"$name$headersKeyValueDelim$value" },
+      mapValue = (session, sessionHeaders) ⇒ sessionHeaders.split(",").toList
     )
 
     def hasSize(expected: Int) = from_session_step(
