+++ src/main/scala/com/github/agourlay/cornichon/http/server/MockServerRequestHandler.scala
@@ -11,7 +11,7 @@ import akka.http.scaladsl.model._
 import akka.http.scaladsl.unmarshalling.Unmarshal
 import akka.stream.ActorMaterializer
 import com.github.agourlay.cornichon.http.{ HttpMethod ⇒ CornichonHttpMethod, HttpMethods ⇒ CornichonHttpMethods, HttpRequest ⇒ CornichonHttpRequest }
-import com.github.agourlay.cornichon.util.ShowInstances._
+import com.github.agourlay.cornichon.util.Instances._
 import com.github.agourlay.cornichon.http.server.MockServerResultsHolder._
 import com.github.agourlay.cornichon.json.CornichonJson
 import io.circe.Json
