+++ src/main/scala/com/github/agourlay/cornichon/steps/regular/assertStep/Diff.scala
@@ -2,7 +2,7 @@ package com.github.agourlay.cornichon.steps.regular.assertStep
 
 import cats.Show
 import com.github.agourlay.cornichon.json.CornichonJson._
-import com.github.agourlay.cornichon.util.ShowInstances._
+import com.github.agourlay.cornichon.util.Instances._
 import cats.syntax.show._
 import io.circe.Json
