+++ src/main/scala/com/github/agourlay/cornichon/steps/wrapped/WithDataInputStep.scala
@@ -9,7 +9,7 @@ import com.github.agourlay.cornichon.json.CornichonJson
 import com.github.agourlay.cornichon.core.Engine._
 import com.github.agourlay.cornichon.core.Done._
 import com.github.agourlay.cornichon.util.Timing._
-import com.github.agourlay.cornichon.util.ShowInstances._
+import com.github.agourlay.cornichon.util.Instances._
 
 import scala.concurrent.{ ExecutionContext, Future }
