+++ src/test/scala/com/github/agourlay/cornichon/json/CornichonJsonSpec.scala
@@ -4,7 +4,7 @@ import io.circe.{ Json, JsonObject }
 import org.scalatest.prop.PropertyChecks
 import org.scalatest.{ Matchers, WordSpec }
 import cats.data.Xor._
-import com.github.agourlay.cornichon.util.ShowInstances._
+import com.github.agourlay.cornichon.util.Instances._
 
 class CornichonJsonSpec extends WordSpec with Matchers with PropertyChecks with CornichonJson {
