+++ src/test/scala/com/github/agourlay/cornichon/steps/StepUtilSpec.scala
@@ -4,11 +4,11 @@ import java.util.Timer
 
 import com.github.agourlay.cornichon.core.Engine
 import com.github.agourlay.cornichon.resolver.Resolver
-import com.github.agourlay.cornichon.util.ShowInstances
+import com.github.agourlay.cornichon.util.Instances
 
 import scala.concurrent.ExecutionContext
 
-trait StepUtilSpec extends ShowInstances {
+trait StepUtilSpec extends Instances {
 
   implicit val ec = ExecutionContext.global
   implicit val timer = new Timer()
